apps/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.38
alembic==1.14.1
psycopg[binary]==3.2.6
redis==5.2.1
pydantic-settings==2.8.1
passlib[argon2]==1.7.4
PyJWT==2.10.1
email-validator==2.2.0
